[ML] Adding index pattern runtime fields to anomaly detection wizards (#91168)
* [ML] Adding index pattern runtime fields to anomaly detection wizards

* hook refactor

* small refactor of search json

* fixing mml estimation error

* changes based on review

* sorting fields in metric selection

* using useMemo rather than useState
