+++ x-pack/plugins/ml/public/application/jobs/new_job/common/chart_loader/searches.ts
@@ -8,6 +8,7 @@
 import { get } from 'lodash';
 
 import { ml } from '../../../../services/ml_api_service';
+import { RuntimeMappings } from '../../../../../../common/types/fields';
 
 interface CategoryResults {
   success: boolean;
@@ -18,7 +19,8 @@ export function getCategoryFields(
   indexPatternName: string,
   fieldName: string,
   size: number,
-  query: any
+  query: any,
+  runtimeMappings?: RuntimeMappings
 ): Promise<CategoryResults> {
   return new Promise((resolve, reject) => {
     ml.esSearch({
@@ -34,6 +36,7 @@ export function getCategoryFields(
             },
           },
         },
+        ...(runtimeMappings !== undefined ? { runtime_mappings: runtimeMappings } : {}),
       },
     })
       .then((resp: any) => {
