+++ x-pack/plugins/ml/public/application/jobs/new_job/common/job_creator/job_creator.ts
@@ -6,10 +6,15 @@
  */
 
 import { BehaviorSubject } from 'rxjs';
+import { cloneDeep } from 'lodash';
 import { SavedSearchSavedObject } from '../../../../../../common/types/kibana';
 import { UrlConfig } from '../../../../../../common/types/custom_urls';
 import { IndexPatternTitle } from '../../../../../../common/types/kibana';
-import { ML_JOB_AGGREGATION } from '../../../../../../common/constants/aggregation_types';
+import {
+  ML_JOB_AGGREGATION,
+  aggregations,
+  mlOnlyAggregations,
+} from '../../../../../../common/constants/aggregation_types';
 import { ES_FIELD_TYPES } from '../../../../../../../../../src/plugins/data/public';
 import {
   Job,
@@ -20,7 +25,8 @@ import {
   BucketSpan,
   CustomSettings,
 } from '../../../../../../common/types/anomaly_detection_jobs';
-import { Aggregation, Field } from '../../../../../../common/types/fields';
+import { Aggregation, Field, RuntimeMappings } from '../../../../../../common/types/fields';
+import { combineFieldsAndAggs } from '../../../../../../common/util/fields_utils';
 import { createEmptyJob, createEmptyDatafeed } from './util/default_configs';
 import { mlJobService } from '../../../../services/job_service';
 import { JobRunner, ProgressSubscriber } from '../job_runner';
@@ -57,7 +63,8 @@ export class JobCreator {
   protected _aggs: Aggregation[] = [];
   protected _fields: Field[] = [];
   protected _scriptFields: Field[] = [];
-  protected _runtimeMappings: Field[] = [];
+  protected _runtimeFields: Field[] = [];
+  protected _runtimeMappings: RuntimeMappings | null = null;
   protected _aggregationFields: Field[] = [];
   protected _sparseData: boolean = false;
   private _stopAllRefreshPolls: {
@@ -85,6 +92,8 @@ export class JobCreator {
     if (typeof indexPattern.timeFieldName === 'string') {
       this._job_config.data_description.time_field = indexPattern.timeFieldName;
     }
+
+    this._extractRuntimeMappings();
 
     this._datafeed_config.query = query;
   }
@@ -489,8 +498,12 @@ export class JobCreator {
     return this._scriptFields;
   }
 
-  public get runtimeMappings(): Field[] {
+  public get runtimeMappings(): RuntimeMappings | null {
     return this._runtimeMappings;
+  }
+
+  public get runtimeFields(): Field[] {
+    return this._runtimeFields;
   }
 
   public get aggregationFields(): Field[] {
@@ -498,7 +511,7 @@ export class JobCreator {
   }
 
   public get additionalFields(): Field[] {
-    return [...this._scriptFields, ...this._runtimeMappings, ...this._aggregationFields];
+    return [...this._scriptFields, ...this._runtimeFields, ...this._aggregationFields];
   }
 
   public get subscribers(): ProgressSubscriber[] {
@@ -662,6 +675,52 @@ export class JobCreator {
     this._job_config.analysis_config.per_partition_categorization!.stop_on_warn = enabled;
   }
 
+  private _extractRuntimeMappings() {
+    const runtimeFieldMap = this._indexPattern.toSpec().runtimeFieldMap;
+    if (runtimeFieldMap !== undefined) {
+      if (this._datafeed_config.runtime_mappings === undefined) {
+        this._datafeed_config.runtime_mappings = {};
+      }
+      Object.entries(runtimeFieldMap).forEach(([key, val]) => {
+        this._datafeed_config.runtime_mappings![key] = val;
+      });
+    }
+    this._populateRuntimeFields();
+  }
+
+  private _populateRuntimeFields() {
+    this._runtimeFields = [];
+    this._runtimeMappings = this._datafeed_config.runtime_mappings ?? null;
+    if (this._runtimeMappings !== null) {
+      const tempRuntimeFields = Object.entries(this._runtimeMappings).map(
+        ([id, runtimeField]) =>
+          ({
+            id,
+            name: id,
+            type: runtimeField.type,
+            aggregatable: true,
+            aggs: [],
+            runtimeField,
+          } as Field)
+      );
+
+      const aggs = cloneDeep([...aggregations, ...mlOnlyAggregations]);
+      this._runtimeFields = combineFieldsAndAggs(tempRuntimeFields, aggs, {}).fields;
+    }
+  }
+
+  private _populateScriptFields() {
+    this._scriptFields = [];
+    if (this._datafeed_config.script_fields !== undefined) {
+      this._scriptFields = Object.keys(this._datafeed_config.script_fields).map((f) => ({
+        id: f,
+        name: f,
+        type: ES_FIELD_TYPES.KEYWORD,
+        aggregatable: true,
+      }));
+    }
+  }
+
   protected _overrideConfigs(job: Job, datafeed: Datafeed) {
     this._job_config = job;
     this._datafeed_config = datafeed;
@@ -683,25 +742,8 @@ export class JobCreator {
       this.useDedicatedIndex = true;
     }
 
-    this._scriptFields = [];
-    if (this._datafeed_config.script_fields !== undefined) {
-      this._scriptFields = Object.keys(this._datafeed_config.script_fields).map((f) => ({
-        id: f,
-        name: f,
-        type: ES_FIELD_TYPES.KEYWORD,
-        aggregatable: true,
-      }));
-    }
-
-    this._runtimeMappings = [];
-    if (this._datafeed_config.runtime_mappings !== undefined) {
-      this._runtimeMappings = Object.keys(this._datafeed_config.runtime_mappings).map((f) => ({
-        id: f,
-        name: f,
-        type: ES_FIELD_TYPES.KEYWORD,
-        aggregatable: true,
-      }));
-    }
+    this._populateScriptFields();
+    this._populateRuntimeFields();
 
     this._aggregationFields = [];
     const aggs = getDatafeedAggregations(this._datafeed_config);
