+++ x-pack/plugins/ml/public/application/jobs/new_job/common/results_loader/results_loader.ts
@@ -255,7 +255,8 @@ export class ResultsLoader {
     if (isMultiMetricJobCreator(this._jobCreator)) {
       if (this._jobCreator.splitField !== null) {
         const fieldValues = await this._chartLoader.loadFieldExampleValues(
-          this._jobCreator.splitField
+          this._jobCreator.splitField,
+          this._jobCreator.runtimeMappings
         );
         if (fieldValues.length > 0) {
           this._detectorSplitFieldFilters = {
