+++ x-pack/plugins/ml/public/application/jobs/new_job/pages/components/pick_fields_step/components/bucket_span_estimator/estimate_bucket_span.ts
@@ -40,6 +40,7 @@ export function useEstimateBucketSpan() {
     query: mlContext.combinedQuery,
     splitField: undefined,
     timeField: mlContext.currentIndexPattern.timeFieldName,
+    runtimeMappings: jobCreator.runtimeMappings ?? undefined,
   };
 
   if (
