+++ x-pack/plugins/ml/public/application/jobs/new_job/pages/components/pick_fields_step/components/multi_metric_view/metric_selection.tsx
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import React, { Fragment, FC, useContext, useEffect, useState } from 'react';
+import React, { Fragment, FC, useContext, useEffect, useState, useMemo } from 'react';
 
 import { JobCreatorContext } from '../../../job_creator_context';
 import { MultiMetricJobCreator } from '../../../../../common/job_creator';
@@ -13,6 +13,7 @@ import { LineChartData } from '../../../../../common/chart_loader';
 import { DropDownLabel, DropDownProps } from '../agg_select';
 import { newJobCapsService } from '../../../../../../../services/new_job_capabilities_service';
 import { AggFieldPair } from '../../../../../../../../../common/types/fields';
+import { sortFields } from '../../../../../../../../../common/util/fields_utils';
 import { getChartSettings, defaultChartSettings } from '../../../charts/common/settings';
 import { MetricSelector } from './metric_selector';
 import { ChartGrid } from './chart_grid';
@@ -33,7 +34,10 @@ export const MultiMetricDetectors: FC<Props> = ({ setIsValid }) => {
 
   const jobCreator = jc as MultiMetricJobCreator;
 
-  const { fields } = newJobCapsService;
+  const fields = useMemo(
+    () => sortFields([...newJobCapsService.fields, ...jobCreator.runtimeFields]),
+    []
+  );
   const [selectedOptions, setSelectedOptions] = useState<DropDownProps>([]);
   const [aggFieldPairList, setAggFieldPairList] = useState<AggFieldPair[]>(
     jobCreator.aggFieldPairs
@@ -107,7 +111,7 @@ export const MultiMetricDetectors: FC<Props> = ({ setIsValid }) => {
   useEffect(() => {
     if (splitField !== null) {
       chartLoader
-        .loadFieldExampleValues(splitField)
+        .loadFieldExampleValues(splitField, jobCreator.runtimeMappings)
         .then(setFieldValues)
         .catch((error) => {
           getToastNotificationService().displayErrorToast(error);
@@ -135,7 +139,8 @@ export const MultiMetricDetectors: FC<Props> = ({ setIsValid }) => {
           aggFieldPairList,
           jobCreator.splitField,
           fieldValues.length > 0 ? fieldValues[0] : null,
-          cs.intervalMs
+          cs.intervalMs,
+          jobCreator.runtimeMappings
         );
         setLineChartsData(resp);
       } catch (error) {
