+++ x-pack/plugins/ml/public/application/jobs/new_job/pages/components/pick_fields_step/components/single_metric_view/metric_selection_summary.tsx
@@ -58,7 +58,8 @@ export const SingleMetricDetectorsSummary: FC = () => {
           [jobCreator.aggFieldPair],
           null,
           null,
-          cs.intervalMs
+          cs.intervalMs,
+          jobCreator.runtimeMappings
         );
         if (resp[DTR_IDX] !== undefined) {
           setLineChartData(resp);
