+++ x-pack/plugins/ml/public/application/jobs/new_job/pages/components/pick_fields_step/components/split_field/by_field.tsx
@@ -5,13 +5,16 @@
  * 2.0.
  */
 
-import React, { FC, useContext, useEffect, useState } from 'react';
+import React, { FC, useContext, useEffect, useState, useMemo } from 'react';
 import { i18n } from '@kbn/i18n';
 
 import { SplitFieldSelect } from './split_field_select';
 import { JobCreatorContext } from '../../../job_creator_context';
 import { Field } from '../../../../../../../../../common/types/fields';
-import { newJobCapsService } from '../../../../../../../services/new_job_capabilities_service';
+import {
+  newJobCapsService,
+  filterCategoryFields,
+} from '../../../../../../../services/new_job_capabilities_service';
 import { MultiMetricJobCreator, PopulationJobCreator } from '../../../../../common/job_creator';
 
 interface Props {
@@ -22,7 +25,11 @@ export const ByFieldSelector: FC<Props> = ({ detectorIndex }) => {
   const { jobCreator: jc, jobCreatorUpdate, jobCreatorUpdated } = useContext(JobCreatorContext);
   const jobCreator = jc as PopulationJobCreator;
 
-  const { categoryFields: allCategoryFields } = newJobCapsService;
+  const runtimeCategoryFields = useMemo(() => filterCategoryFields(jobCreator.runtimeFields), []);
+  const allCategoryFields = useMemo(
+    () => [...newJobCapsService.categoryFields, ...runtimeCategoryFields],
+    []
+  );
 
   const [byField, setByField] = useState(jobCreator.getByField(detectorIndex));
   const categoryFields = useFilteredCategoryFields(
