+++ x-pack/plugins/ml/public/application/jobs/new_job/pages/components/pick_fields_step/components/split_field/split_field.tsx
@@ -5,11 +5,14 @@
  * 2.0.
  */
 
-import React, { FC, useContext, useEffect, useState } from 'react';
+import React, { FC, useContext, useEffect, useState, useMemo } from 'react';
 
 import { SplitFieldSelect } from './split_field_select';
 import { JobCreatorContext } from '../../../job_creator_context';
-import { newJobCapsService } from '../../../../../../../services/new_job_capabilities_service';
+import {
+  newJobCapsService,
+  filterCategoryFields,
+} from '../../../../../../../services/new_job_capabilities_service';
 import { Description } from './description';
 import {
   MultiMetricJobCreator,
@@ -23,7 +26,11 @@ export const SplitFieldSelector: FC = () => {
   const jobCreator = jc as MultiMetricJobCreator | PopulationJobCreator;
   const canClearSelection = isMultiMetricJobCreator(jc);
 
-  const { categoryFields } = newJobCapsService;
+  const runtimeCategoryFields = useMemo(() => filterCategoryFields(jobCreator.runtimeFields), []);
+  const categoryFields = useMemo(
+    () => [...newJobCapsService.categoryFields, ...runtimeCategoryFields],
+    []
+  );
   const [splitField, setSplitField] = useState(jobCreator.splitField);
 
   useEffect(() => {
