+++ x-pack/plugins/ml/server/models/bucket_span_estimator/bucket_span_estimator.test.ts
@@ -35,6 +35,7 @@ const formConfig: BucketSpanEstimatorData = {
   },
   splitField: undefined,
   timeField: undefined,
+  runtimeMappings: undefined,
 };
 
 describe('ML - BucketSpanEstimator', () => {
