+++ x-pack/plugins/ml/server/models/job_service/new_job/categorization/examples.ts
@@ -14,6 +14,7 @@ import {
   CategorizationAnalyzer,
   CategoryFieldExample,
 } from '../../../../../common/types/categories';
+import { RuntimeMappings } from '../../../../../common/types/fields';
 import { ValidationResults } from './validation_results';
 
 const CHUNK_SIZE = 100;
@@ -32,7 +33,8 @@ export function categorizationExamplesProvider({
     timeField: string | undefined,
     start: number,
     end: number,
-    analyzer: CategorizationAnalyzer
+    analyzer: CategorizationAnalyzer,
+    runtimeMappings: RuntimeMappings | undefined
   ): Promise<{ examples: CategoryFieldExample[]; error?: any }> {
     if (timeField !== undefined) {
       const range = {
@@ -65,6 +67,7 @@ export function categorizationExamplesProvider({
         _source: false,
         query,
         sort: ['_doc'],
+        ...(runtimeMappings !== undefined ? { runtime_mappings: runtimeMappings } : {}),
       },
     });
 
@@ -165,7 +168,8 @@ export function categorizationExamplesProvider({
     timeField: string | undefined,
     start: number,
     end: number,
-    analyzer: CategorizationAnalyzer
+    analyzer: CategorizationAnalyzer,
+    runtimeMappings: RuntimeMappings | undefined
   ) {
     const resp = await categorizationExamples(
       indexPatternTitle,
@@ -175,7 +179,8 @@ export function categorizationExamplesProvider({
       timeField,
       start,
       end,
-      analyzer
+      analyzer,
+      runtimeMappings
     );
 
     const { examples } = resp;
