+++ x-pack/plugins/ml/server/models/job_service/new_job/line_chart.ts
@@ -7,7 +7,11 @@
 
 import { get } from 'lodash';
 import { IScopedClusterClient } from 'kibana/server';
-import { AggFieldNamePair, EVENT_RATE_FIELD_ID } from '../../../../common/types/fields';
+import {
+  AggFieldNamePair,
+  EVENT_RATE_FIELD_ID,
+  RuntimeMappings,
+} from '../../../../common/types/fields';
 import { ML_MEDIAN_PERCENTS } from '../../../../common/util/job_utils';
 
 type DtrIndex = number;
@@ -34,7 +38,8 @@ export function newJobLineChartProvider({ asCurrentUser }: IScopedClusterClient)
     query: object,
     aggFieldNamePairs: AggFieldNamePair[],
     splitFieldName: string | null,
-    splitFieldValue: string | null
+    splitFieldValue: string | null,
+    runtimeMappings: RuntimeMappings | undefined
   ) {
     const json: object = getSearchJsonFromConfig(
       indexPatternTitle,
@@ -45,7 +50,8 @@ export function newJobLineChartProvider({ asCurrentUser }: IScopedClusterClient)
       query,
       aggFieldNamePairs,
       splitFieldName,
-      splitFieldValue
+      splitFieldValue,
+      runtimeMappings
     );
 
     const { body } = await asCurrentUser.search(json);
@@ -103,7 +109,8 @@ function getSearchJsonFromConfig(
   query: any,
   aggFieldNamePairs: AggFieldNamePair[],
   splitFieldName: string | null,
-  splitFieldValue: string | null
+  splitFieldValue: string | null,
+  runtimeMappings: RuntimeMappings | undefined
 ): object {
   const json = {
     index: indexPatternTitle,
@@ -125,6 +132,7 @@ function getSearchJsonFromConfig(
           aggs: {},
         },
       },
+      ...(runtimeMappings !== undefined ? { runtime_mappings: runtimeMappings } : {}),
     },
   };
