+++ x-pack/plugins/ml/server/routes/schemas/job_service_schema.ts
@@ -16,6 +16,7 @@ export const categorizationFieldExamplesSchema = {
   start: schema.number(),
   end: schema.number(),
   analyzer: schema.any(),
+  runtimeMappings: schema.maybe(schema.any()),
 };
 
 export const chartSchema = {
@@ -28,6 +29,7 @@ export const chartSchema = {
   aggFieldNamePairs: schema.arrayOf(schema.any()),
   splitFieldName: schema.maybe(schema.nullable(schema.string())),
   splitFieldValue: schema.maybe(schema.nullable(schema.string())),
+  runtimeMappings: schema.maybe(schema.any()),
 };
 
 export const datafeedIdsSchema = schema.object({
